Maitred Transaction Monitor Version 1.0 Upload
Version 1.0
Initial Program is very stiff. Everything is hard coded and not easily
configured. It is setup for 13 terminals starting at: 192.168.100.101 -
192.168.100.113, Sending data to port 5555 at 192.168.100.114.

The Function of this program is to accept packets on port 5555 and
manage the data in a separate thread per connection. Each thread checks
the IP of the connected address and saves the data in an appropriate log
file.